Optionally discard chains (dflt. no)
We used to always discard chains before when working with the
surface graph, but when using the contacts graph we can't discard
them. We *could* if we also kept track of which chains are
removed and remove them from the graph as well. I'm not sure if
that can disconnect the graph, uh, it would be something to check
if we did that in the future.