ThesisPipeline to speed up experiment design
Most of my models follow the same pipeline with little changes.
The idea behind this class is to provide boxes and by calling
each method the boxes are connected, thus completing an experiment.
In essence the state of the notebook is replaced by the state of
the class. This cleans up a lot of the visual clutter, allows
for easy experiment design and execution and it is more mantainable.
Later I plan on incorporating the XGBoost model into this class,
and creating an Executor class that reports the configuration
and runs the experiments.